+++ app/Models/Simrs/TindakanSimrs.php
@@ -29,7 +29,7 @@ class TindakanSimrs extends Model
 
     public function getTotalBiayaAttribute()
     {
-        $total = (int) round($this->jumlah * round($this->tindakan_biaya));
+        $total = round($this->jumlah * round($this->tindakan_biaya));
         $discount = round($total * ($this->discount / 100));
         $total_biaya = $total - $discount;
         return $total_biaya;
